+++ Src/Library/changelog.md
@@ -1,6 +1,7 @@
 ### NEW
 - jwt refresh token support
 - export swagger json file with dotnet run/ms build #348
+- specify binding sources for the default request binder [#info](https://discord.com/channels/933662816458645504/1045775010226253876)
 
 ### IMPROVEMENTS
 - tighten up even handling to support long-running processes